style: clean up #35
