--- src/main/java/com/kospot/kospot/application/game/roadView/rank/EndRoadViewRankUseCase.java
@@ -1,48 +1,0 @@
-package com.kospot.kospot.application.game.roadView.rank;
-
-import com.kospot.kospot.domain.game.dto.request.EndGameRequest;
-import com.kospot.kospot.domain.game.dto.response.EndGameResponse;
-import com.kospot.kospot.domain.game.entity.GameType;
-import com.kospot.kospot.domain.game.entity.RoadViewGame;
-import com.kospot.kospot.domain.game.service.RoadViewGameService;
-import com.kospot.kospot.domain.gameRank.adaptor.GameRankAdaptor;
-import com.kospot.kospot.domain.gameRank.entity.GameRank;
-import com.kospot.kospot.domain.gameRank.service.GameRankService;
-import com.kospot.kospot.domain.member.entity.Member;
-import com.kospot.kospot.domain.point.entity.PointHistoryType;
-import com.kospot.kospot.domain.point.service.PointHistoryService;
-import com.kospot.kospot.domain.point.service.PointService;
-import com.kospot.kospot.global.annotation.usecase.UseCase;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-
-@UseCase
-@RequiredArgsConstructor
-@Transactional
-public class EndRoadViewRankUseCase {
-
-    private final RoadViewGameService roadViewGameService;
-    private final PointService pointService;
-    private final GameRankAdaptor gameRankAdaptor;
-    private final PointHistoryService pointHistoryService;
-    private final GameRankService gameRankService;
-
-    //todo refactor transaction
-    public EndGameResponse.RoadViewRank execute(Member member, EndGameRequest.RoadView request){
-        // end
-        RoadViewGame game = roadViewGameService.endRankGame(member, request);
-
-        // earn point
-        GameRank gameRank = gameRankAdaptor.queryByMemberAndGameType(member, GameType.ROADVIEW);
-        int point = pointService.addPointByRankGameScore(member, gameRank, game.getScore());
-
-        // calculate rating point
-        gameRankService.updateRatingScoreAfterGameEnd(gameRank, game);
-
-        // save point history
-        pointHistoryService.savePointHistory(member, point, PointHistoryType.RANK_GAME);
-
-        return EndGameResponse.RoadViewRank.from(game);
-    }
-
-}
