+++ src/main/java/com/kospot/kospot/application/game/roadView/rank/usecase/EndRoadViewRankUseCaseV2.java
@@ -1,4 +1,4 @@
-package com.kospot.kospot.application.game.roadView.rank;
+package com.kospot.kospot.application.game.roadView.rank.usecase;
 
 import com.kospot.kospot.domain.game.entity.GameType;
 import com.kospot.kospot.domain.game.event.RoadViewGameEvent;
@@ -10,7 +10,6 @@ import com.kospot.kospot.domain.gameRank.adaptor.GameRankAdaptor;
 import com.kospot.kospot.domain.gameRank.entity.GameRank;
 import com.kospot.kospot.domain.member.entity.Member;
 import com.kospot.kospot.global.annotation.usecase.UseCase;
-import jakarta.persistence.EntityManager;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
