+++ src/main/java/com/kospot/kospot/presentation/game/RoadViewGameController.java
@@ -2,8 +2,8 @@ package com.kospot.kospot.presentation.game;
 
 import com.kospot.kospot.application.game.roadView.practice.EndRoadViewPracticeUseCase;
 import com.kospot.kospot.application.game.roadView.practice.StartRoadViewPracticeUseCase;
-import com.kospot.kospot.application.game.roadView.rank.EndRoadViewRankUseCase;
-import com.kospot.kospot.application.game.roadView.rank.StartRoadViewRankUseCase;
+import com.kospot.kospot.application.game.roadView.rank.usecase.EndRoadViewRankUseCase;
+import com.kospot.kospot.application.game.roadView.rank.usecase.StartRoadViewRankUseCase;
 import com.kospot.kospot.domain.game.dto.request.EndGameRequest;
 import com.kospot.kospot.domain.game.dto.response.EndGameResponse;
 import com.kospot.kospot.domain.game.dto.response.StartGameResponse;
