+++ src/test/java/com/kospot/kospot/game/service/RoadViewGameListenerTest.java
@@ -1,13 +1,12 @@
 package com.kospot.kospot.game.service;
 
-import com.kospot.kospot.application.game.roadView.rank.EndRoadViewRankUseCase;
-import com.kospot.kospot.application.game.roadView.rank.EndRoadViewRankUseCaseV2;
+import com.kospot.kospot.application.game.roadView.rank.usecase.EndRoadViewRankUseCase;
+import com.kospot.kospot.application.game.roadView.rank.usecase.EndRoadViewRankUseCaseV2;
 import com.kospot.kospot.application.game.roadView.rank.listener.EndRoadViewRankEventListener;
 import com.kospot.kospot.domain.game.dto.request.EndGameRequest;
 import com.kospot.kospot.domain.game.dto.response.EndGameResponse;
 import com.kospot.kospot.domain.game.entity.GameType;
 import com.kospot.kospot.domain.game.entity.RoadViewGame;
-import com.kospot.kospot.domain.game.event.RoadViewGameEvent;
 import com.kospot.kospot.domain.game.repository.RoadViewGameRepository;
 import com.kospot.kospot.domain.game.service.RoadViewGameService;
 import com.kospot.kospot.domain.gameRank.entity.GameRank;
@@ -17,6 +16,7 @@ import com.kospot.kospot.domain.member.entity.Member;
 import com.kospot.kospot.domain.member.repository.MemberRepository;
 import com.kospot.kospot.domain.point.adaptor.PointHistoryAdaptor;
 import com.kospot.kospot.domain.point.entity.PointHistory;
+import com.kospot.kospot.domain.point.repository.PointHistoryRepository;
 import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -26,18 +26,14 @@ import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabas
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.test.context.event.RecordApplicationEvents;
-import org.springframework.transaction.annotation.Transactional;
 
-import java.sql.Time;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.Future;
-import java.util.concurrent.TimeUnit;
 
 
 import static org.awaitility.Awaitility.await;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.mockito.ArgumentMatchers.any;
 
 @SpringBootTest
@@ -65,6 +61,9 @@ public class RoadViewGameListenerTest {
 
     @Autowired
     private PointHistoryAdaptor pointHistoryAdaptor;
+
+    @Autowired
+    private PointHistoryRepository pointHistoryRepository;
 
     @Autowired
     private EndRoadViewRankEventListener endRoadViewRankEventListener;
@@ -112,43 +111,30 @@ public class RoadViewGameListenerTest {
     @Test
     @DisplayName("기존 게임 종료 로직 시간 측정")
     void testGameEnd() {
-        // given
+        //given
         long startTime = System.currentTimeMillis();
-
-        // when
-        List<EndGameResponse.RoadViewRank> responses = new ArrayList<>();
-
 
         //when
         for (int i = 0; i < 100; i++) {
             Member member = members.get(i);
-            responses.add(endRoadViewRankUseCase.execute(member, requests.get(i)));
+            endRoadViewRankUseCase.execute(member, requests.get(i));
 
         }
-
         // then
         long endTime = System.currentTimeMillis();
         System.out.println("기존 게임 종료 로직 시간: " + (endTime - startTime) + "ms");
 
-        for ( int i = 0 ; i < 100; i++) {
-            Member member = members.get(i);
-            System.out.println("gameId: " + games.get(i).getId());
-            System.out.println("gameRanks.get(i).getRatingScore(): " + gameRanks.get(i).getRatingScore());
-            System.out.println("responses.get(i).getratingPointChange(): " + responses.get(i).getRatingScoreChange());
-            System.out.println("games.get(i).getScore(): " + games.get(i).getScore());
-            int ratingScoreChange = RatingScoreCalculator.calculateRatingChange(games.get(i).getScore(), 0);
-            System.out.println("RatingScoreCalculator.calculateRatingChange(games.get(i).getScore(), gameRanks.get(i).getRatingScore()): " + ratingScoreChange);
-            assertEquals(ratingScoreChange, responses.get(i).getRatingScoreChange());
+        //persist
+        Member member = memberRepository.findById(members.get(0).getId()).orElseThrow();
 
-            //point history
-            List<PointHistory> pointHistories = pointHistoryAdaptor.queryAllHistoryByMemberId(member.getId());
-            System.out.println("point history change amount: " + pointHistories.get(0).getChangeAmount());
-            System.out.println("point history " + pointHistories.get(0).getPointHistoryType());
+        int countPointHistory = (int) pointHistoryRepository.count();
+        assertEquals(100, countPointHistory);
 
-            //member point
-            assertEquals(pointHistories.get(0).getChangeAmount(), member.getPoint());
+        int memberPoint = member.getPoint();
+        assertNotEquals(0, memberPoint);
 
-        }
+        int ratingScore = gameRankRepository.findByMemberAndGameType(member, GameType.ROADVIEW).getRatingScore();
+        assertNotEquals(0, ratingScore);
     }
 
     @Test
@@ -156,38 +142,30 @@ public class RoadViewGameListenerTest {
     void testGameEndEventListener() throws InterruptedException {
         //given
         long startTime = System.currentTimeMillis();
-        List<EndGameResponse.RoadViewRank> responses = new ArrayList<>();
+
         //when
         for (int i = 0; i < 100; i++) {
-            responses.add(endRoadViewRankUseCaseV2.execute(members.get(i), requests.get(i)));
+            Member member = members.get(i);
+            endRoadViewRankUseCaseV2.execute(member, requests.get(i));
         }
 
         //then
         long endTime = System.currentTimeMillis();
         System.out.println("게임 종료 이벤트 리스너 시간: " + (endTime - startTime) + "ms");
 
-//        Thread.sleep(5000);
+        Thread.sleep(1000);
 
-        for ( int i = 0 ; i < 100; i++) {
-            Member member = members.get(i);
+        //persist
+        Member member = memberRepository.findById(members.get(0).getId()).orElseThrow();
 
-            System.out.println("gameId: " + games.get(i).getId());
-            System.out.println("gameRanks.get(i).getRatingScore(): " + gameRanks.get(i).getRatingScore());
-            System.out.println("responses.get(i).getratingPointChange(): " + responses.get(i).getRatingScoreChange());
-            System.out.println("games.get(i).getScore(): " + games.get(i).getScore());
-            int ratingScoreChange = RatingScoreCalculator.calculateRatingChange(games.get(i).getScore(), 0);
-            System.out.println("RatingScoreCalculator.calculateRatingChange(games.get(i).getScore(), gameRanks.get(i).getRatingScore()): " + ratingScoreChange);
-//            assertEquals(ratingScoreChange, responses.get(i).getRatingScoreChange());
+        int countPointHistory = (int) pointHistoryRepository.count();
+        assertEquals(100, countPointHistory);
 
-            //point history
-//            List<PointHistory> pointHistories = pointHistoryAdaptor.queryAllHistoryByMemberId(member.getId());
-//            System.out.println("point history change amount: " + pointHistories.get(0).getChangeAmount());
-//            System.out.println("point history " + pointHistories.get(0).getPointHistoryType());
-//
-//            //member point
-//            assertEquals(pointHistories.get(0).getChangeAmount(), member.getPoint());
-        }
+        int memberPoint = member.getPoint();
+        assertNotEquals(0, memberPoint);
 
+        int ratingScore = gameRankRepository.findByMemberAndGameType(member, GameType.ROADVIEW).getRatingScore();
+        assertNotEquals(0, ratingScore);
 
     }
 }
