+++ src/test/java/com/kospot/kospot/game/service/RoadViewGameServiceTest.java
@@ -1,7 +1,7 @@
 package com.kospot.kospot.game.service;
 
 import com.kospot.kospot.application.game.roadView.practice.EndRoadViewPracticeUseCase;
-import com.kospot.kospot.application.game.roadView.rank.EndRoadViewRankUseCase;
+import com.kospot.kospot.application.game.roadView.rank.usecase.EndRoadViewRankUseCase;
 import com.kospot.kospot.domain.game.dto.request.EndGameRequest;
 import com.kospot.kospot.domain.game.dto.response.EndGameResponse;
 import com.kospot.kospot.domain.game.entity.GameMode;
